dm-3508
Map and info window styles.. enums for Diffusion History statuses.. etc.
