+++ app/assets/stylesheets/dm/components/_diffusion_history.scss
@@ -7,8 +7,6 @@
 
 .community_map_styles{
   height: 540px;
-  width: 100%;
-  max-width: 726px;
 }
 
 .home-marker-modal {
