+++ app/assets/stylesheets/dm/pages/_practice.scss
@@ -990,21 +990,12 @@ li.small-disc:before {
 
 .map-infowindow-size{
   width: 575px !important;
-  min-width: 200px !important;
+  //min-width: 200px !important;
   height: 339px !important;
 }
-.map-infowindow-border{
-  width: 560px !important;
-  min-width: 200px !important;
-  border-width: 0.01rem !important;
-  border-color: lightgray !important;
-  border-style: dotted !important;
-  border-radius: 4px !important;
-  padding: 2% 3% 2% 4% !important;
-}
-.padding-for-map-component-infowindow-text{
-  padding-left: 4% !important;
-  display:block !important;
+
+.gm-style-iw.gm-style-iw-c{
+  padding: 56px 8px 52px 8px;
 }
 
 tr.border_bottom td {
