+++ app/views/page/_page_map.html.erb
@@ -5,6 +5,6 @@
     <% end %>
     <%= javascript_include_tag 'page/page_map', 'data-turbolinks-track': 'reload' %>
   <% end %>
-  <div id="page_builder_map" class="display-none community_map_styles"></div>
+  <div id="page_builder_map" class="display-none grid-col-12 desktop:grid-col-7 community_map_styles"></div>
   <%= render partial: 'shared/loading_spinner', locals: { display: true, classes: 'flex-justify-center flex-align-self-center margin-y-8 desktop:margin-y-10 dm-facilities-show-map-loading-spinner text-center' } %>
 <% end %>
